feat: restructure asset folder; adjust contentful & sitemap script
