+++ .github/workflows/cron.yml
@@ -8,18 +8,19 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-        with:
-          ref: master
       - name: Setup node
         uses: actions/setup-node@v4
         with:
           node-version: 20
       - name: Fetch Contentful data
         run: SPACE_ID=${{secrets.CONTENTFUL_SPACE_ID}} ACCESS_TOKEN=${{secrets.CONTENTFUL_ACCESS_TOKEN}} node ./contentful.js
+      - name: Update sitemap
+        run: node ./sitemap.js
       - name: Commit to GitHub
         run: |
           git config --global user.name "GitHub Actions"
           git config --global user.email "actions@github.com"
-          git add contentfulEntries.json sitemap.txt
+          git add app/_assets/contentfulEntries.json
+          git add public/sitemap.txt
           git commit -m "Update Contentful data & sitemap" || echo "Nothing to commit"
           git push --no-verify
