+++ app/components/cover/cover.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import styled from 'styled-components'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import coverImg from '@public/images/cover.jpg'
+import coverImg from '@assets/images/cover.jpg'
 import { SOCIAL_MEDIA_LIST, SocialMediaInfo } from '@/constants'
 
 const CoverBackground = styled.div`
