--- legacy/public/sitemap.txt
@@ -1,1 +1,0 @@
-https://pov.tw
